+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Bank Reconciliation: Default toDate value is now filled.
 - DEMA DATA : fix display only english and french at Import selection
 - Base : Resolve Email PDF printing issue.
+- Stock Move mass invoicing: correct context on generated invoice.
 
 ## [5.1.1] - 2019-01-18
 ## Features